add ur5e launch and config templates;
should still be tested on the cell
