+++ moveit2_workshop_bringup/config/app/ur5e_cell/app_marker.yaml
@@ -1,0 +1,5 @@
+/**:
+  ros__parameters:
+    planning_group: "manipulator" # All target poses are planned w.r.t the base_link frame defined for this group in the SRDF.
+    marker1_frame: "marker_1"
+    marker2_frame: "marker_2"
