+++ moveit2_workshop_bringup/config/app/ur5e_cell/app_simple.yaml
@@ -1,0 +1,5 @@
+/**:
+  ros__parameters:
+    planning_group: "manipulator" # All target poses are planned w.r.t the base_link frame defined for this group in the SRDF.
+    target1_pose:  [0.5, 0.2, 0.2, 1.0, 0.0, 0.0, 0.0 ] # geometry_msgs/msg/Pose position and orientation values
+    target2_pose: [0.5, -0.2, 0.2, 1.0, 0.0, 0.0, 0.0 ] # geometry_msgs/msg/Pose position and orientation values
